// pages/contentTypes.d.ts
export interface ApiPublicationPublication extends Schema.CollectionType {
  collectionName: 'publications';
  info: {
    singularName: 'publication';
    pluralName: 'publications';
    displayName: 'Publication';
    description: '';
  };
  options: {
    draftAndPublish: true;
  };
  attributes: {
    Title: Attribute.String;
    PublicationDate: Attribute.Date;
    Authors: Attribute.Text;
    Slug: Attribute.UID<'api::publication.publication', 'Title'>;
    seo: Attribute.Component<'shared.seo'>;
    Image: Attribute.Media & Attribute.Required;
    Abstract: Attribute.RichText & Attribute.Required;
    ReadMoreUrl: Attribute.String & Attribute.Required;
    createdAt: Attribute.DateTime;
    updatedAt: Attribute.DateTime;
    publishedAt: Attribute.DateTime;
    createdBy: Attribute.Relation<
      'api::publication.publication',
      'oneToOne',
      'admin::user'
    > &
      Attribute.Private;
    updatedBy: Attribute.Relation<
      'api::publication.publication',
      'oneToOne',
      'admin::user'
    > &
      Attribute.Private;
  };
}